Update bolt bracing instructions
The explanation of bolt bracing in the readme was confusing to me,

> Holding the bolt back on a gun and letting go of the foregrip and grip while shouldering the gun will keep it in place, freeing up one hand while leaving the bolt open

And what I interpreted this to mean is that I should pull the bolt back and then let both grips go on both hands ("letting go of the foregrip and grip") which would make the gun return to my harnessed quickslot ("while shouldering the gun"). And that this shouldn't shoulder the gun, but should instead keep the gun in one of my hands ("freeing up one hand") so that I can use the other hand to top it up.

It was only when I saw one of the other dot points that it made sense:

> Works on Bolt Action rifles, Closed Bolt guns, and handguns ***which have a stock***

Hopefully this new description will stop others from getting confused?